Feat: Rename BuildRunner into DockerPipelineRunner
